Added a timestamp comparison for the CHECKSUMS.md5 file.
The idea here is to detect if someone is trying to provide us with an
older version of the CHECKSUMS.md5 file. That is, trying to keep us from
updating to the latest patches.

An attacker with a MiTM position, might want to keep us from updating,
or even make it so, that we downgrade to older (vulnerable) versions.

Got the idea from this Tor ticket:
https://trac.torproject.org/projects/tor/ticket/2340