Created a global variable QUIET_GPGV
This way we can control the gpgv quiet mode more efficiently (vs.
the -q switch to gpg_verify()).

The idea is that you can see the CHECKSUMS.md5 verified in the
beginning, but not every time it's called from md5_verify() during the
MD5 checks of the actual packages... so that the package upgrade output
wouldn't get so cluttered.